test: update test fixtures
